fix: Address review feedback — early-return paths, warn on failures, test file placement
- Add ensure_auto_compact_settings() to early-return paths in all 3 worktree
  creation functions so pre-existing worktrees get the settings file written
- Replace silent `let _ =` with tracing::warn! on write failures
- Move auto_compact tests to separate settings_tests.rs per project convention
