+++ src/settings.rs
@@ -180,13 +180,26 @@ pub fn ensure_auto_compact_settings(dir: &std::path::Path) {
         // File exists but missing autoCompact — add it
         let mut obj = existing;
         obj["autoCompact"] = serde_json::Value::Bool(true);
-        let _ = std::fs::write(&settings_path, serde_json::to_string_pretty(&obj).unwrap());
+        if let Err(e) = std::fs::write(&settings_path, serde_json::to_string_pretty(&obj).unwrap())
+        {
+            tracing::warn!(
+                "Failed to write autoCompact to {}: {}",
+                settings_path.display(),
+                e
+            );
+        }
         return;
     }
 
     // File doesn't exist or isn't valid JSON — write a minimal one
-    if std::fs::create_dir_all(&claude_dir).is_ok() {
-        let _ = std::fs::write(&settings_path, "{\n  \"autoCompact\": true\n}\n");
+    if std::fs::create_dir_all(&claude_dir).is_ok()
+        && let Err(e) = std::fs::write(&settings_path, "{\n  \"autoCompact\": true\n}\n")
+    {
+        tracing::warn!(
+            "Failed to create autoCompact settings at {}: {}",
+            settings_path.display(),
+            e
+        );
     }
 }
 
@@ -285,66 +298,8 @@ mod tests {
             "settings should not contain unreplaced {{bin}} placeholders"
         );
     }
-
-    /// `ensure_auto_compact_settings` creates `.claude/settings.json` with autoCompact.
-    ///
-    /// Fork sessions don't receive `--settings <file>` (they use `--resume --fork-session`
-    /// which skips the explicit settings file to avoid duplicate tool registrations).
-    /// However, ALL sessions get `--setting-sources project,local`, so they read
-    /// `.claude/settings.json`. Writing this file into worktrees is the project-agnostic
-    /// mechanism for fork sessions to get `autoCompact: true`.
-    ///
-    /// Regression test for !2177 / !2180.
-    #[test]
-    fn test_ensure_auto_compact_creates_settings() {
-        let dir = tempfile::TempDir::new().expect("temp dir");
-
-        // File doesn't exist yet
-        let settings_path = dir.path().join(".claude/settings.json");
-        assert!(!settings_path.exists());
-
-        // Call should create it
-        ensure_auto_compact_settings(dir.path());
-        assert!(settings_path.exists());
-        let content = std::fs::read_to_string(&settings_path).unwrap();
-        let settings: serde_json::Value = serde_json::from_str(&content).unwrap();
-        assert_eq!(settings["autoCompact"], true);
-    }
-
-    #[test]
-    fn test_ensure_auto_compact_idempotent() {
-        let dir = tempfile::TempDir::new().expect("temp dir");
-
-        // Create the file first
-        ensure_auto_compact_settings(dir.path());
-
-        // Call again — should not error
-        ensure_auto_compact_settings(dir.path());
-
-        let settings_path = dir.path().join(".claude/settings.json");
-        let content = std::fs::read_to_string(&settings_path).unwrap();
-        let settings: serde_json::Value = serde_json::from_str(&content).unwrap();
-        assert_eq!(settings["autoCompact"], true);
-    }
-
-    #[test]
-    fn test_ensure_auto_compact_preserves_existing_keys() {
-        let dir = tempfile::TempDir::new().expect("temp dir");
-        let claude_dir = dir.path().join(".claude");
-        std::fs::create_dir_all(&claude_dir).unwrap();
-
-        // Write a settings file with other keys but no autoCompact
-        std::fs::write(
-            claude_dir.join("settings.json"),
-            r#"{"someKey": "someValue"}"#,
-        )
-        .unwrap();
-
-        ensure_auto_compact_settings(dir.path());
-
-        let content = std::fs::read_to_string(claude_dir.join("settings.json")).unwrap();
-        let settings: serde_json::Value = serde_json::from_str(&content).unwrap();
-        assert_eq!(settings["autoCompact"], true);
-        assert_eq!(settings["someKey"], "someValue");
-    }
 }
+
+#[path = "settings_tests.rs"]
+#[cfg(test)]
+mod auto_compact_tests;
